backend/requirements.txt
Django==2.2.27
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
drf-base64==2.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.20.0
